+++ database/migrations/2016_11_19_111648_AddProviderID.php
@@ -13,10 +13,17 @@ class AddProviderID extends Migration
      */
     public function up()
     {
-        Schema::table('users', function ($table) {
-            $table->dropColumn('facebook_id');
-            $table->dropColumn('google_id');
-        });
+        if(Schema::hasColumn('users','facebook_id')){
+            Schema::table('users', function ($table) {
+                $table->dropColumn('facebook_id');
+            });
+        }
+
+        if(Schema::hasColumn('users','google_id')){
+            Schema::table('users', function ($table) {
+                $table->dropColumn('google_id');
+            });
+        }
 
         Schema::table('users', function ($table) {
             $table->string('social_provider')->nullable()->after('remember_token');
